+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,6 +1,7 @@
 ﻿### Version 3.8.0
 
 - Improvement: Added support for Razor file code management.
+- Improvement: Added `CSharpCollectionExpression` to the C# File Builder which can be used to generate C# [collection expressions](https://learn.microsoft.com/dotnet/csharp/language-reference/operators/collection-expressions).
 - Fixed: Nullable aware assignments of instantiated objects via mapping systems not resolving replacements correctly, leading to uncompilable code.
 - Fixed: Mapping system unable to find constructor when multiple partial C# files are in play.